@@ -1,5 +1,3 @@
-module github.com/mattn/go-runewidth
+module zgo.at/runewidth
 
-go 1.9
-
-require github.com/rivo/uniseg v0.2.0
+go 1.16
